Screenshots fix
