+++ src/test/java/ApplicationRunTest.java
@@ -1,17 +1,17 @@
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
+import io.cucumber.testng.AbstractTestNGCucumberTests;
+import io.cucumber.testng.CucumberOptions;
+import org.testng.annotations.DataProvider;
 @SuppressWarnings("unused")
-@RunWith(Cucumber.class)
 @CucumberOptions(
         features = "src/test/resources/features",
         glue = "stepDefs",
         tags = "@avito"
 )
-public class ApplicationRunTest {
-    @Test
-    public static void run(String[] args) {
+public class ApplicationRunTest extends AbstractTestNGCucumberTests {
+
+    @Override
+    @DataProvider
+    public Object[][] scenarios() {
+        return super.scenarios();
     }
 }
