+++ src/test/java/stepDefs/enumiration/Filters.java
@@ -1,16 +1,19 @@
-package enumiration;
+package stepDefs.enumiration;
 
 import lombok.Getter;
 
 @Getter
 @SuppressWarnings({"NonAsciiCharacters", "unused"})
-public enum Categories {
-    оргтехника(0, "Оргтехника и расходники");
+public enum Filters {
+
+    По_умолчанию(0, "По умолчанию"),
+    Дороже(1, "Дороже"),
+    Дешевле(2,"Дешевле"),
+    По_дате(3,"По дате");
 
     private final int id;
     private final String name;
-
-    Categories(int id, String name) {
+    Filters(int id, String name){
         this.id = id;
         this.name = name;
     }
