+++ src/test/resources/features/Avito.feature
@@ -18,4 +18,4 @@
     Примеры:
       | техника | город           | количество | фильтр  |
       | принтер | Владивосток     | 3          | Дороже  |
-      | сканер  | Санкт-Петербург | 5          | Дороже |
+      | сканер  | Санкт-Петербург | 5          | Дешевле |
